patch history commits
