+++ connector_dns/__init__.py
@@ -1,9 +1,3 @@
 # -*- coding: utf-8 -*-
-# Copyright 2015 Elico Corp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import backend
-from . import connector
+from . import components
 from . import models
-from . import unit
-
