+++ connector_dns/__manifest__.py
