--- connector_dns/backend.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2015 Elico Corp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-import openerp.addons.connector.backend as backend
-
-dns = backend.Backend('dns')
