--- connector_dns/data/dns_record_type.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data noupdate="1">
-
-    <record id="type_a" model="dns.record.type">
-        <field name="name">IPv4 Address</field>
-        <field name="code">A</field>
-        <field name="help">IPv4 address. Enter multiple addresses 
-                               on separate lines.
-                           Example: 
-                               192.0.2.235 
-                               198.51.100.234
-        </field>
-        <field name="validate_regex">^((?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\n?)+$</field>
-    </record>
-    
-    <record id="type_aaaa" model="dns.record.type">
-        <field name="name">IPv6 Address</field>
-        <field name="code">AAAA</field>
-        <field name="help">IPv6 address. Enter multiple addresses 
-                               on separate lines.
-                           Example: 
-                               2001:0db8:85a3:0:0:8a2e:0370:7334 
-                               fe80:0:0:0:202:b3ff:fe1e:8329
-        </field>
-        <field name="validate_regex">^((?:[A-F0-9]{1,4}:){7}[A-F0-9]{1,4}\n?)+$</field>
-    </record>
-    
-    <record id="type_cname" model="dns.record.type">
-        <field name="name">Canonical Name</field>
-        <field name="code">CNAME</field>
-        <field name="help">The domain name that you want to 
-                               resolve to instead of the value in the 
-                               Name field.
-                           Example: 
-                               www.example.com
-        </field>
-        <field name="validate_regex">^([a-z0-9]+(-[a-z0-9]+)*\.)+[a-z]{2,}$</field>
-    </record>
-    
-    <record id="type_mx" model="dns.record.type">
-        <field name="name">Mail Exchange</field>
-        <field name="code">MX</field>
-        <field name="help">A priority and a domain name that 
-                               specifies a mail server. Enter multiple 
-                               values on separate lines.
-                           Format: 
-                               [priority] [mail server host name] 
-                           Example: 
-                               10 mailserver.example.com. 
-                               20 mailserver2.example.com.
-        </field>
-        <field name="validate_regex">^([12]?\d ([a-z0-9]+(-[a-z0-9]+)*\.)+[a-z]{2,}\n)+$</field>
-    </record>
-    
-    <record id="type_txt" model="dns.record.type">
-        <field name="name">Text</field>
-        <field name="code">TXT</field>
-        <field name="help">A text record.
-                               Enter multiple values on separate lines.
-                               Enclose text in quotation marks.
-                           Example: 
-                                "Sample Text Entries" 
-                                "Enclose entries in quotation marks"
-        </field>
-        <field name="validate_regex">^(".*"\n)+$</field>
-    </record>
-    
-    <record id="type_ptr" model="dns.record.type">
-        <field name="name">Pointer</field>
-        <field name="code">PTR</field>
-        <field name="help">The domain name that you want to return.
-                           Example: 
-                               www.example.com
-        </field>
-        <field name="validate_regex">^([a-z0-9]+(-[a-z0-9]+)*\.)+[a-z]{2,}$</field>
-    </record>
-    
-    <record id="type_srv" model="dns.record.type">
-        <field name="name">Service Locator</field>
-        <field name="code">SRV</field>
-        <field name="help">A SRV record. For information about SRV 
-                               record format, refer to the applicable 
-                               documentation. Enter multiple values 
-                               on separate lines.
-                           Format: 
-                               [priority] [weight] [port] [server host name]
-                           Example: 
-                               1 10 5269 xmpp-server.example.com. 
-                               2 12 5060 sip-server.example.com.
-        </field>
-        <field name="validate_regex">^(\d+ \d+ \d+ ([a-z0-9]+(-[a-z0-9]+)*\.)+[a-z]{2,}\n)+$</field>
-    </record>
-    
-    <record id="type_spf" model="dns.record.type">
-        <field name="name">Sender Policy Framework</field>
-        <field name="code">SPF</field>
-        <field name="help">An SPF record. For information about SPF 
-                               record format, refer to the applicable 
-                               documentation. Enter multiple values 
-                               on separate lines. Enclose values in 
-                               quotation marks.
-                           Example: 
-                               "v=spf1 ip4:192.168.0.1/16-all"
-        </field>
-        <field name="validate_regex">.*</field>
-    </record>
-    
-    <record id="type_naptr" model="dns.record.type">
-        <field name="name">Name Authority Pointer</field>
-        <field name="code">NAPTR</field>
-        <field name="help">An NAPTR record. For information about NAPTR 
-                               record format, refer to the applicable 
-                               documentation. Enter multiple values 
-                               on separate lines.
-                           Format: 
-                               [order] [preference] [flags] [services] [regexp] [replacement]
-                           Example: 
-                               100 100 "U" "" "!^.*$!sip:info@bar.example.com!" . 
-                               10 100 "S" "SIP+D2U" "" foo.example.com.
-        </field>
-        <field name="validate_regex">.*</field>
-    </record>
-    
-    <record id="type_ns" model="dns.record.type">
-        <field name="name">Name Server</field>
-        <field name="code">A</field>
-        <field name="help">The domain name of a name server. 
-                               Enter multiple name servers on 
-                               separate lines.
-                           Example: 
-                               ns1.amazon.com 
-                               ns2.amazon.org 
-                               ns3.amazon.net 
-                               ns4.amazon.co.uk
-        </field>
-        <field name="validate_regex">^(([a-z0-9]+(-[a-z0-9]+)*\.)+[a-z]{2,}\n)+$</field>
-    </record>
-
-</data>
-</openerp>
