+++ connector_dns/models/__init__.py
@@ -1,8 +1,4 @@
-# -*- coding: utf-8 -*-
-# Copyright 2015 Elico Corp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from . import dns_backend
-from . import dns_zone
+from . import dns_binding
+from . import dns_domain
 from . import dns_record
-from . import dns_record_type
