--- connector_dns/security/dns.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-
-        <record id="group_dns_user" model="res.groups">
-            <field name="name">DNS User</field>
-            <field name="users" eval="[(4, ref('base.user_root'))]" />
-            <field name="implied_ids" eval="[(4, ref('base.group_user'))]" />
-        </record>
-        
-        <record id="group_dns_manager" model="res.groups">
-            <field name="name">DNS Manager</field>
-            <field name="users" eval="[(4, ref('base.user_root'))]" />
-            <field name="implied_ids" eval="[(4, ref('base.group_user')),
-                                             (4, ref('group_dns_user'))]"
-                   />
-        </record>
-
-    </data>
-</openerp>
