--- connector_dns/tests/__init__.py
@@ -1,32 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2015 Elico Corp
-# Copyright 2016 LasLabs Inc.
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-# Common
-from . import common
-
-# Unit
-from . import test_binder
-from . import test_connector
-from . import test_consumer
-from . import test_backend_adapter
-
-# Mapper
-from . import test_dns_import_mapper
-
-# Importer
-from . import test_batch_importer
-from . import test_delayed_batch_importer
-from . import test_direct_batch_importer
-from . import test_dns_importer
-
-# Exporter
-from . import test_base_exporter
-from . import test_dns_exporter
-
-# Deleter
-from . import test_dns_deleter
-
-# Models
-from .models import *
