--- connector_dns/tests/models/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2016 LasLabs Inc.
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import test_dns_backend
-from . import test_dns_record
-from . import test_dns_record_type
