--- connector_dns/unit/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2015 Elico Corp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import backend_adapter
-from . import binder
-from . import export_synchronizer
-from . import import_synchronizer
-from . import delete_synchronizer
-from . import mapper
