--- connector_dns/views/dns_backend.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        
-        <record id="dns_backend_action" model="ir.actions.act_window">
-            <field name="name">DNS Backends</field>
-            <field name="res_model">dns.backend</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <record id="dns_backend_tree" model="ir.ui.view">
-            <field name="name">DNS Backend Tree</field>
-            <field name="model">dns.backend</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="login" />
-                </tree>
-            </field>
-        </record>
-
-        <record id="dns_backend_form" model="ir.ui.view">
-            <field name="name">DNS Backend Form</field>
-            <field name="model">dns.backend</field>
-            <field name="arch" type="xml">
-                <form>
-                    <header>
-                        <button name="button_connect"
-                                string="Connect"
-                                type="object"
-                                />
-                        <button name="button_set_draft"
-                                string="Set to draft"
-                                type="object"
-                                />
-                    </header>
-                    <sheet>
-                        <div class="oe_title" style="width: 390px;">
-                            <label class="oe_edit_only" for="name" string="Name" />
-                            <h1>
-                                <field name="name" class="oe_inline" />
-                            </h1>
-                        </div>
-                        <group name="connector">
-                            <field name="version" />
-                            <field name="company_id" />
-                            <field name="is_default" />
-                            <field name="active" />
-                        </group>
-                        <group name="credentials">
-                            <field name="uri" />
-                            <field name="login" widget="email" />
-                            <field name="password" password="True" />
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        
-    </data>
-</openerp>
