--- connector_dns/views/dns_record.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        
-        <record id="dns_record_action" model="ir.actions.act_window">
-            <field name="name">DNS Records</field>
-            <field name="res_model">dns.record</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <record id="dns_record_tree" model="ir.ui.view">
-            <field name="name">DNS Record Tree</field>
-            <field name="model">dns.record</field>
-            <field name="arch" type="xml">
-                <tree editable="top">
-                    <field name="name" />
-                    <field name="zone_id" />
-                    <field name="type_id" />
-                    <field name="value" />
-                    <field name="ttl" />
-                </tree>
-            </field>
-        </record>
-
-        <record id="dns_record_form" model="ir.ui.view">
-            <field name="name">DNS Record Form</field>
-            <field name="model">dns.record</field>
-            <field name="arch" type="xml">
-                <form>
-                    <header />
-                    <sheet>
-                        <div class="oe_title" style="width: 390px;">
-                            <label class="oe_edit_only" for="name" string="Record Name" />
-                            <h1>
-                                <field name="name" class="oe_inline" />
-                            </h1>
-                        </div>
-                        <group>
-                            <field name="zone_id" />
-                            <field name="type_id" />
-                            <field name="ttl" />
-                            <field name="value" />
-                            <field name="dns_bind_ids" />
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        
-    </data>
-</openerp>
