--- dns/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import components
-from . import models
