--- dns/models/__init__.py
@@ -1,4 +1,0 @@
-from . import dns_backend
-from . import dns_binding
-from . import dns_domain
-from . import dns_record
